wallet connect action card into actionbuilder
